main/requirement.txt
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
cloudpickle==3.0.0
distro==1.8.0
eli5==0.13.0
exceptiongroup==1.2.0
filelock==3.13.1
fsspec==2023.12.2
graphviz==0.20.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.1
idna==3.6
imbalanced-learn==0.12.0
importlib-metadata==7.0.1
Jinja2==3.1.2
joblib==1.3.2
llvmlite==0.41.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
openai==1.6.0
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.10.1
shap==0.44.0
six==1.16.0
sklearn-contrib-lightning==0.6.2.post0
slicer==0.0.7
sniffio==1.3.0
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
zipp==3.17.0
